+++ packages/language/src/format.ts
@@ -372,7 +372,7 @@ export function formatExpression(
     case "Ternary": {
       const n = node as Ternary
       // HubL supports a JS-style ternary (`cond ? a : b`) in addition to the standard
-      // HubL/Jinja conditional expression (`a if cond else b`). Preserve the operator style
+      // conditional expression (`a if cond else b`). Preserve the operator style
       // if it was parsed from `? :`.
       const isQMarkTernary = n.ifToken?.value === "?"