+++ packages/server/src/definition.ts
@@ -29,7 +29,7 @@ export const getDefinition = async (uri: string, position: lsp.Position) => {
     | undefined
 
   // Handle namespaced macro calls, e.g.
-  // `{% import "lib.hubl" as lib %}` then `{{ lib.example() }}`.
+  // `{% import "lib.html" as lib %}` then `{{ lib.example() }}`.
   // In this case the call callee is a MemberExpression, not an Identifier.
   if (
     callExpression !== undefined &&